\name{print.flexCrossHaz}
\alias{print.flexCrossHaz}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for the flexCrossHaz class }
\description{
  Printing the most important feautures of an estimated crossing point.
}
\usage{
\method{print}{flexCrossHaz}(x, digits = max(3, getOption("digits") - 3), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x }{ object of class \code{flexCrossHaz} }
  \item{digits }{ number of digits to be printed }
  \item{\dots }{ arguments passed to other functions }
}
\author{ Vito M.R. Muggeo }

\keyword{ survival }